Refactor `MassEncryption` into the core entity
