+++ app/jobs/mass_encryption/application_job.rb
@@ -1,4 +1,4 @@
-module MassEncryption
+class MassEncryption
   class ApplicationJob < ActiveJob::Base
     queue_as :encryption
   end
