+++ lib/mass_encryption/version.rb
@@ -1,3 +1,3 @@
-module MassEncryption
+class MassEncryption
   VERSION = '0.1.1'
 end
