+++ test/mass_encryption_test.rb
@@ -1,16 +1,16 @@
 require "test_helper"
 
-class EncryptorTest < ActiveSupport::TestCase
+class MassEncryptionTest < ActiveSupport::TestCase
   test "encrypt all the records in parallel" do
     perform_enqueued_jobs only: MassEncryption::BatchEncryptionJob do
-      MassEncryption::Encryptor.new.encrypt_all_later
+      MassEncryption.new.perform_later
     end
     assert_everything_is_encrypted
   end
 
   test "encrypt all the records in tracks" do
     perform_enqueued_jobs only: MassEncryption::BatchEncryptionJob do
-      MassEncryption::Encryptor.new(tracks_count: 4, batch_size: 2).encrypt_all_later
+      MassEncryption.new(tracks_count: 4, batch_size: 2).perform_later
     end
     assert_everything_is_encrypted
   end
@@ -20,7 +20,7 @@ class EncryptorTest < ActiveSupport::TestCase
     first_post_to_encrypt = all_posts.all[5]
 
     perform_enqueued_jobs only: MassEncryption::BatchEncryptionJob do
-      MassEncryption::Encryptor.new(from_id: first_post_to_encrypt.id).encrypt_all_later
+      MassEncryption.new(from_id: first_post_to_encrypt.id).perform_later
     end
 
     assert_encrypted_records all_posts.where("id >= ?", first_post_to_encrypt.id)
@@ -32,7 +32,7 @@ class EncryptorTest < ActiveSupport::TestCase
     first_post_to_encrypt = all_posts.all[5]
 
     perform_enqueued_jobs only: MassEncryption::BatchEncryptionJob do
-      MassEncryption::Encryptor.new(from_id: first_post_to_encrypt.id, tracks_count: 4, batch_size: 2).encrypt_all_later
+      MassEncryption.new(from_id: first_post_to_encrypt.id, tracks_count: 4, batch_size: 2).perform_later
     end
 
     assert_encrypted_records all_posts.where("id >= ?", first_post_to_encrypt.id)
@@ -41,7 +41,7 @@ class EncryptorTest < ActiveSupport::TestCase
 
   test "encrypting in tracks create the expected tracks" do
     assert_enqueued_jobs 2, only: MassEncryption::BatchEncryptionJob do
-      MassEncryption::Encryptor.new(only: Post, tracks_count: 2, batch_size: 2).encrypt_all_later
+      MassEncryption.new(only: Post, tracks_count: 2, batch_size: 2).perform_later
     end
 
     batch_1, batch_2 = enqueued_jobs.collect { |serialized_job| instantiate_job(serialized_job).arguments.first }.flatten
@@ -52,7 +52,7 @@ class EncryptorTest < ActiveSupport::TestCase
 
   test "encrypting in tracks won't encrypt things more than once" do
     assert_enqueued_jobs 2, only: MassEncryption::BatchEncryptionJob do
-      MassEncryption::Encryptor.new(only: Post, tracks_count: 2, batch_size: Post.count / 4).encrypt_all_later
+      MassEncryption.new(only: Post, tracks_count: 2, batch_size: Post.count / 4).perform_later
     end
 
     batch_1_1, batch_1_2 = enqueued_jobs.collect { |serialized_job| instantiate_job(serialized_job).arguments.first }.flatten
@@ -74,7 +74,7 @@ class EncryptorTest < ActiveSupport::TestCase
 
   test "provide classes to encrypt" do
     perform_enqueued_jobs only: MassEncryption::BatchEncryptionJob do
-      MassEncryption::Encryptor.new(only: [ Person ]).encrypt_all_later
+      MassEncryption.new(only: [ Person ]).perform_later
     end
 
     assert_not_encrypted_records Post.all
@@ -84,7 +84,7 @@ class EncryptorTest < ActiveSupport::TestCase
 
   test "exclude records to encrypt" do
     perform_enqueued_jobs only: MassEncryption::BatchEncryptionJob do
-      MassEncryption::Encryptor.new(except: [ Person ]).encrypt_all_later
+      MassEncryption.new(except: [ Person ]).perform_later
     end
 
     assert_encrypted_records Post.all
@@ -95,11 +95,11 @@ class EncryptorTest < ActiveSupport::TestCase
   test "when running in tracks, it enqueues successive jobs until the whole batch is encrypted" do
     assert Post.count > 1
 
-    MassEncryption::Encryptor.new(only: [ Post ], batch_size: 1, tracks_count: 1).encrypt_all_later
+    MassEncryption.new(only: [ Post ], batch_size: 1, tracks_count: 1).perform_later
     assert_enqueued_jobs 1, only: MassEncryption::BatchEncryptionJob
 
     perform_enqueued_jobs only: MassEncryption::BatchEncryptionJob do
-      MassEncryption::Encryptor.new(only: [ Post ], batch_size: 1).encrypt_all_later
+      MassEncryption.new(only: [ Post ], batch_size: 1).perform_later
     end
 
     assert_performed_jobs Post.count, only: MassEncryption::BatchEncryptionJob
@@ -108,11 +108,11 @@ class EncryptorTest < ActiveSupport::TestCase
   test "when running in parallel, it enqueues all the needed jobs to encrypt the batch" do
     assert Post.count > 1
 
-    MassEncryption::Encryptor.new(only: [ Post ], batch_size: 1).encrypt_all_later
+    MassEncryption.new(only: [ Post ], batch_size: 1).perform_later
     assert_enqueued_jobs Post.count, only: MassEncryption::BatchEncryptionJob
 
     perform_enqueued_jobs only: MassEncryption::BatchEncryptionJob do
-      MassEncryption::Encryptor.new(only: [ Post ], batch_size: 1).encrypt_all_later
+      MassEncryption.new(only: [ Post ], batch_size: 1).perform_later
     end
 
     assert_performed_jobs Post.count, only: MassEncryption::BatchEncryptionJob
@@ -120,7 +120,7 @@ class EncryptorTest < ActiveSupport::TestCase
 
   test "encrypting includes encrypted rich texts attributes" do
     perform_enqueued_jobs only: MassEncryption::BatchEncryptionJob do
-      MassEncryption::Encryptor.new.encrypt_all_later
+      MassEncryption.new.perform_later
     end
 
     assert_encrypted_records ActionText::EncryptedRichText.all
@@ -131,13 +131,13 @@ class EncryptorTest < ActiveSupport::TestCase
 
     assert_nothing_raised do
       perform_enqueued_jobs only: MassEncryption::BatchEncryptionJob do
-        MassEncryption::Encryptor.new.encrypt_all_later
+        MassEncryption.new.perform_later
       end
     end
 
     assert_nothing_raised do
       perform_enqueued_jobs only: MassEncryption::BatchEncryptionJob do
-        MassEncryption::Encryptor.new(tracks_count: 1).encrypt_all_later
+        MassEncryption.new(tracks_count: 1).perform_later
       end
     end
   end
